fix: address adversarial review R4 findings for gym achievements
- Move badge aggregate recomputation to session-start (always, not gated)
- Add common badge aggregate update + checkCommonAchievements to victory lock
- Fix duplicate reward notification to not falsely claim new pokemon
